+++ python/CHANGELOG.md
@@ -14,6 +14,7 @@ semver guidelines for more details about this.
 
 - Mark python 3.13 as supported.
 - Bugfix: limit the number of bytes we read in case of an input with just many whitespaces. ([#1015](https://github.com/google/magika/pull/1015))
+- Add `asdict()` utility method to `MagikaResult`.
 
 
 ## [0.6.1] - 2025-03-19